requirements.txt
numpy
opencv-python
Pillow
PyYAML
tqdm
scipy
google-genai
requests
